Relocated calculate_dihedral_angle from conformers to common
And changed the `coord` argument to `coords`
